fix batch issue seen on screen session
These lines make sure that root is started in batch mode, no matter
which one is the first module to be pulled in.  Without these lines I
observe (but not with all analyses) the error below:

    X connection to localhost:32.0 broken (explicit kill or server shutdown).

For some reason, which I don't understand, this error seems to show up
only when running supy within a screen session (i.e. it doesn't show
up with the same analysis running within ssh). However, I can
reproduce this error message on lxplus and on other machines (most of
them running SLC5).