gethostname -> getfqdn
Details:
Some of the batch info is picked up from the hostname.
On the new slc6 nodes, `gethostname` does not provide the domain, so
the job submission will fail (see below, then it falls back on the
`default` site).
I don't know whether this is an ATLAS-specific problem, but this
solution shouldn't have any drawback.
```
@lxplus0487 supy $ python --version
Python 2.6.6
@lxplus0487 supy $ python -c 'import socket; print socket.getfqdn(); print socket.gethostname()'
lxplus0487.cern.ch
lxplus0487
```